+++ FirebaseCore/Tests/Unit/FIRAppTest.m
@@ -32,6 +32,7 @@ NSString *const kFIRTestAppName2 = @"test-app-name-2";
 - (BOOL)configureCore;
 + (NSError *)errorForInvalidAppID;
 - (BOOL)isAppIDValid;
+- (BOOL)isGlobalDataCollectionEnabled;
 + (NSString *)actualBundleID;
 + (NSNumber *)mapFromServiceStringToTypeEnum:(NSString *)serviceString;
 + (NSString *)deviceModel;
@@ -582,6 +583,8 @@ NSString *const kFIRTestAppName2 = @"test-app-name-2";
 
 #pragma mark - Automatic Data Collection Tests
 
+#pragma mark Reading Plist and User Defaults Combinations
+
 - (void)testGlobalDataCollectionNoFlags {
   // Test: No flags set.
   NSString *name = NSStringFromSelector(_cmd);
@@ -592,7 +595,14 @@ NSString *const kFIRTestAppName2 = @"test-app-name-2";
   OCMStub([self.appClassMock readDataCollectionSwitchFromUserDefaultsForApp:OCMOCK_ANY])
       .andReturn(nil);
 
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
   XCTAssertTrue(app.isDataCollectionDefaultEnabled);
+#pragma clang diagnostic pop
+
+  // Test the internal flag and the state matches.
+  XCTAssertTrue(app.isGlobalDataCollectionEnabled);
+  XCTAssertEqual(app.dataCollectionDefaultState, FIRDataCollectionStateDefault);
 }
 
 - (void)testGlobalDataCollectionPlistSetEnabled {
@@ -605,7 +615,16 @@ NSString *const kFIRTestAppName2 = @"test-app-name-2";
   OCMStub([self.appClassMock readDataCollectionSwitchFromUserDefaultsForApp:OCMOCK_ANY])
       .andReturn(nil);
 
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
   XCTAssertTrue(app.isDataCollectionDefaultEnabled);
+#pragma clang diagnostic pop
+
+  // Test the internal flag and the state matches.
+  XCTAssertTrue(app.isGlobalDataCollectionEnabled);
+
+  // Note: this is still `default` because it's the runtime flag. It wasn't explicitly set.
+  XCTAssertEqual(app.dataCollectionDefaultState, FIRDataCollectionStateDefault);
 }
 
 - (void)testGlobalDataCollectionPlistSetDisabled {
@@ -618,7 +637,16 @@ NSString *const kFIRTestAppName2 = @"test-app-name-2";
   OCMStub([self.appClassMock readDataCollectionSwitchFromUserDefaultsForApp:OCMOCK_ANY])
       .andReturn(nil);
 
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
   XCTAssertFalse(app.isDataCollectionDefaultEnabled);
+#pragma clang diagnostic pop
+
+  // Test the internal flag and the state matches.
+  XCTAssertFalse(app.isGlobalDataCollectionEnabled);
+
+  // Note: this is still `default` because it's the runtime flag. It wasn't explicitly set.
+  XCTAssertEqual(app.dataCollectionDefaultState, FIRDataCollectionStateDefault);
 }
 
 - (void)testGlobalDataCollectionUserSpecifiedEnabled {
@@ -631,7 +659,14 @@ NSString *const kFIRTestAppName2 = @"test-app-name-2";
   OCMStub([self.appClassMock readDataCollectionSwitchFromUserDefaultsForApp:OCMOCK_ANY])
       .andReturn(@YES);
 
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
   XCTAssertTrue(app.isDataCollectionDefaultEnabled);
+#pragma clang diagnostic pop
+
+  // Test the internal flag and the state matches.
+  XCTAssertTrue(app.isGlobalDataCollectionEnabled);
+  XCTAssertEqual(app.dataCollectionDefaultState, FIRDataCollectionStateEnabled);
 }
 
 - (void)testGlobalDataCollectionUserSpecifiedDisabled {
@@ -644,7 +679,14 @@ NSString *const kFIRTestAppName2 = @"test-app-name-2";
   OCMStub([self.appClassMock readDataCollectionSwitchFromUserDefaultsForApp:OCMOCK_ANY])
       .andReturn(@NO);
 
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
   XCTAssertFalse(app.isDataCollectionDefaultEnabled);
+#pragma clang diagnostic pop
+
+  // Test the internal flag and the state matches.
+  XCTAssertFalse(app.isGlobalDataCollectionEnabled);
+  XCTAssertEqual(app.dataCollectionDefaultState, FIRDataCollectionStateDisabled);
 }
 
 - (void)testGlobalDataCollectionUserOverriddenEnabled {
@@ -657,7 +699,14 @@ NSString *const kFIRTestAppName2 = @"test-app-name-2";
   OCMStub([self.appClassMock readDataCollectionSwitchFromUserDefaultsForApp:OCMOCK_ANY])
       .andReturn(@YES);
 
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
   XCTAssertTrue(app.isDataCollectionDefaultEnabled);
+#pragma clang diagnostic pop
+
+  // Test the internal flag and the state matches.
+  XCTAssertTrue(app.isGlobalDataCollectionEnabled);
+  XCTAssertEqual(app.dataCollectionDefaultState, FIRDataCollectionStateEnabled);
 }
 
 - (void)testGlobalDataCollectionUserOverriddenDisabled {
@@ -670,36 +719,51 @@ NSString *const kFIRTestAppName2 = @"test-app-name-2";
   OCMStub([self.appClassMock readDataCollectionSwitchFromUserDefaultsForApp:OCMOCK_ANY])
       .andReturn(@NO);
 
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
   XCTAssertFalse(app.isDataCollectionDefaultEnabled);
+#pragma clang diagnostic pop
+
+  // Test the internal flag and the state matches.
+  XCTAssertFalse(app.isGlobalDataCollectionEnabled);
+  XCTAssertEqual(app.dataCollectionDefaultState, FIRDataCollectionStateDisabled);
 }
 
-- (void)testGlobalDataCollectionWriteToDefaults {
+#pragma mark Enum: Plist and User Defaults Cominations
+
+- (void)testGlobalDataCollectionEnumWriteToDefaults {
   id defaultsMock = OCMPartialMock([NSUserDefaults standardUserDefaults]);
   NSString *name = NSStringFromSelector(_cmd);
   FIROptions *options = [[FIROptions alloc] initWithGoogleAppID:kGoogleAppID
                                                     GCMSenderID:kGCMSenderID];
   [FIRApp configureWithName:name options:options];
   FIRApp *app = [FIRApp appNamed:name];
-  app.dataCollectionDefaultEnabled = YES;
+
+  app.dataCollectionDefaultState = FIRDataCollectionStateEnabled;
   NSString *key =
       [NSString stringWithFormat:kFIRGlobalAppDataCollectionEnabledDefaultsKeyFormat, app.name];
   OCMVerify([defaultsMock setObject:@YES forKey:key]);
 
-  app.dataCollectionDefaultEnabled = NO;
+  app.dataCollectionDefaultState = FIRDataCollectionStateDisabled;
+
   OCMVerify([defaultsMock setObject:@NO forKey:key]);
 
   [defaultsMock stopMocking];
 }
 
-- (void)testGlobalDataCollectionClearedAfterDelete {
+- (void)testGlobalDataCollectionEnumClearedAfterDelete {
   // Configure and disable data collection for the default FIRApp.
   NSString *name = NSStringFromSelector(_cmd);
   FIROptions *options = [[FIROptions alloc] initWithGoogleAppID:kGoogleAppID
                                                     GCMSenderID:kGCMSenderID];
   [FIRApp configureWithName:name options:options];
   FIRApp *app = [FIRApp appNamed:name];
-  app.dataCollectionDefaultEnabled = NO;
-  XCTAssertFalse(app.isDataCollectionDefaultEnabled);
+
+  app.dataCollectionDefaultState = FIRDataCollectionStateDisabled;
+
+  // Test the internal flag and the state matches.
+  XCTAssertFalse(app.isGlobalDataCollectionEnabled);
+  XCTAssertEqual(app.dataCollectionDefaultState, FIRDataCollectionStateDisabled);
 
   // Delete the app, and verify that the switch was reset.
   XCTestExpectation *deleteFinished =
@@ -714,14 +778,20 @@ NSString *const kFIRTestAppName2 = @"test-app-name-2";
 
   // Set up an app with the same name again, and check the data collection flag.
   [FIRApp configureWithName:name options:options];
-  XCTAssertTrue([FIRApp appNamed:name].isDataCollectionDefaultEnabled);
+
+  // Test the internal flag and the state matches.
+  XCTAssertTrue(app.isGlobalDataCollectionEnabled);
+  XCTAssertEqual(app.dataCollectionDefaultState, FIRDataCollectionStateDefault);
 }
 
-- (void)testGlobalDataCollectionNoDiagnosticsSent {
+- (void)testGlobalDataCollectionEnumNoDiagnosticsSent {
   FIROptions *options = [[FIROptions alloc] initWithGoogleAppID:kGoogleAppID
                                                     GCMSenderID:kGCMSenderID];
   FIRApp *app = [[FIRApp alloc] initInstanceWithName:NSStringFromSelector(_cmd) options:options];
-  app.dataCollectionDefaultEnabled = NO;
+  app.dataCollectionDefaultState = FIRDataCollectionStateDisabled;
+
+  XCTAssertFalse(app.isGlobalDataCollectionEnabled);
+  XCTAssertEqual(app.dataCollectionDefaultState, FIRDataCollectionStateDisabled);
 
   // Add an observer for the diagnostics notification. Currently no object is sent, but in the
   // future that could change so leave it as OCMOCK_ANY.
@@ -743,9 +813,110 @@ NSString *const kFIRTestAppName2 = @"test-app-name-2";
   OCMVerifyAll(self.observerMock);
 }
 
-#pragma mark - Analytics Flag Tests
+#pragma mark Deprecated Flag: Plist and User Defaults Combinations
 
-- (void)testAnalyticsSetByGlobalDataCollectionSwitch {
+- (void)testGlobalDataCollectionWriteToDefaults {
+  id defaultsMock = OCMPartialMock([NSUserDefaults standardUserDefaults]);
+  NSString *name = NSStringFromSelector(_cmd);
+  FIROptions *options = [[FIROptions alloc] initWithGoogleAppID:kGoogleAppID
+                                                    GCMSenderID:kGCMSenderID];
+  [FIRApp configureWithName:name options:options];
+  FIRApp *app = [FIRApp appNamed:name];
+
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
+  app.dataCollectionDefaultEnabled = YES;
+  NSString *key =
+      [NSString stringWithFormat:kFIRGlobalAppDataCollectionEnabledDefaultsKeyFormat, app.name];
+  OCMVerify([defaultsMock setObject:@YES forKey:key]);
+
+  app.dataCollectionDefaultEnabled = NO;
+#pragma clang diagnostic pop
+
+  OCMVerify([defaultsMock setObject:@NO forKey:key]);
+
+  [defaultsMock stopMocking];
+}
+
+- (void)testGlobalDataCollectionClearedAfterDelete {
+  // Configure and disable data collection for the default FIRApp.
+  NSString *name = NSStringFromSelector(_cmd);
+  FIROptions *options = [[FIROptions alloc] initWithGoogleAppID:kGoogleAppID
+                                                    GCMSenderID:kGCMSenderID];
+  [FIRApp configureWithName:name options:options];
+  FIRApp *app = [FIRApp appNamed:name];
+
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
+  app.dataCollectionDefaultEnabled = NO;
+  XCTAssertFalse(app.isDataCollectionDefaultEnabled);
+#pragma clang diagnostic pop
+
+  // Test the internal flag and the state matches.
+  XCTAssertFalse(app.isGlobalDataCollectionEnabled);
+  XCTAssertEqual(app.dataCollectionDefaultState, FIRDataCollectionStateDisabled);
+
+  // Delete the app, and verify that the switch was reset.
+  XCTestExpectation *deleteFinished =
+      [self expectationWithDescription:@"The app should successfully delete."];
+  [app deleteApp:^(BOOL success) {
+    XCTAssertTrue(success);
+    [deleteFinished fulfill];
+  }];
+
+  // Wait for the delete to complete.
+  [self waitForExpectations:@[ deleteFinished ] timeout:1];
+
+  // Set up an app with the same name again, and check the data collection flag.
+  [FIRApp configureWithName:name options:options];
+
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
+  XCTAssertTrue([FIRApp appNamed:name].isDataCollectionDefaultEnabled);
+#pragma clang diagnostic pop
+
+  // Test the internal flag and the state matches.
+  XCTAssertTrue(app.isGlobalDataCollectionEnabled);
+  XCTAssertEqual(app.dataCollectionDefaultState, FIRDataCollectionStateDefault);
+}
+
+- (void)testGlobalDataCollectionBoolNoDiagnosticsSent {
+  FIROptions *options = [[FIROptions alloc] initWithGoogleAppID:kGoogleAppID
+                                                    GCMSenderID:kGCMSenderID];
+  FIRApp *app = [[FIRApp alloc] initInstanceWithName:NSStringFromSelector(_cmd) options:options];
+
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
+  app.dataCollectionDefaultEnabled = NO;
+#pragma clang diagnostic pop
+
+  // Test the internal flag and the state matches.
+  XCTAssertFalse(app.isGlobalDataCollectionEnabled);
+  XCTAssertEqual(app.dataCollectionDefaultState, FIRDataCollectionStateDisabled);
+
+  // Add an observer for the diagnostics notification. Currently no object is sent, but in the
+  // future that could change so leave it as OCMOCK_ANY.
+  [self.notificationCenter addMockObserver:self.observerMock
+                                      name:kFIRAppDiagnosticsNotification
+                                    object:OCMOCK_ANY];
+
+  // Stub out reading from user defaults since stubbing out the BOOL has issues. If the data
+  // collection switch is disabled, the `sendLogs` call should return immediately and not fire a
+  // notification.
+  OCMStub([self.appClassMock readDataCollectionSwitchFromUserDefaultsForApp:OCMOCK_ANY])
+      .andReturn(@NO);
+
+  // Ensure configure doesn't fire a notification.
+  [FIRApp configure];
+
+  // The observer mock is strict and will raise an exception when an unexpected notification is
+  // received.
+  OCMVerifyAll(self.observerMock);
+}
+
+#pragma mark Deprecated Flag: Analytics Flag Tests
+
+- (void)testAnalyticsSetByGlobalDataCollectionBoolSwitch {
   // Test that the global data collection switch triggers setting Analytics when no explicit flag is
   // set.
   id optionsMock = OCMClassMock([FIROptions class]);
@@ -757,15 +928,18 @@ NSString *const kFIRTestAppName2 = @"test-app-name-2";
   id configurationMock = OCMClassMock([FIRAnalyticsConfiguration class]);
   OCMStub([configurationMock sharedInstance]).andReturn(configurationMock);
 
-  // Ensure Analytics is set after the global flag is set. It needs to
+// Ensure Analytics is set after the global flag is set.
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
   [defaultApp setDataCollectionDefaultEnabled:YES];
   OCMVerify([configurationMock setAnalyticsCollectionEnabled:YES persistSetting:NO]);
 
   [defaultApp setDataCollectionDefaultEnabled:NO];
+#pragma clang diagnostic pop
   OCMVerify([configurationMock setAnalyticsCollectionEnabled:NO persistSetting:NO]);
 }
 
-- (void)testAnalyticsNotSetByGlobalDataCollectionSwitch {
+- (void)testAnalyticsNotSetByGlobalDataCollectionBoolSwitch {
   // Test that the global data collection switch doesn't override an explicitly set Analytics flag.
   id optionsMock = OCMClassMock([FIROptions class]);
   OCMStub([optionsMock isAnalyticsCollectionExplicitlySet]).andReturn(YES);
@@ -777,11 +951,39 @@ NSString *const kFIRTestAppName2 = @"test-app-name-2";
   // Reject any changes to Analytics when the data collection changes.
   OCMReject([configurationMock setAnalyticsCollectionEnabled:YES persistSetting:YES]);
   OCMReject([configurationMock setAnalyticsCollectionEnabled:YES persistSetting:NO]);
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
   [app setDataCollectionDefaultEnabled:YES];
 
   OCMReject([configurationMock setAnalyticsCollectionEnabled:NO persistSetting:YES]);
   OCMReject([configurationMock setAnalyticsCollectionEnabled:NO persistSetting:NO]);
   [app setDataCollectionDefaultEnabled:NO];
+#pragma clang diagnostic pop
+}
+
+#pragma mark Enum: Analytics Flag Tests
+
+- (void)testAnalyticsSetByGlobalDataCollectionEnumSwitch {
+  // Test that the global data collection switch triggers setting Analytics when no explicit flag is
+  // set.
+  id optionsMock = OCMClassMock([FIROptions class]);
+  OCMStub([optionsMock isAnalyticsCollectionExplicitlySet]).andReturn(NO);
+
+  // We need to use the default app name since Analytics only associates with the default app.
+  FIRApp *defaultApp = [[FIRApp alloc] initInstanceWithName:kFIRDefaultAppName options:optionsMock];
+
+  id configurationMock = OCMClassMock([FIRAnalyticsConfiguration class]);
+  OCMStub([configurationMock sharedInstance]).andReturn(configurationMock);
+
+  // Ensure Analytics is set after the global flag is set.
+  defaultApp.dataCollectionDefaultState = FIRDataCollectionStateEnabled;
+  OCMVerify([configurationMock setAnalyticsCollectionEnabled:YES persistSetting:NO]);
+
+  defaultApp.dataCollectionDefaultState = FIRDataCollectionStateDisabled;
+  OCMVerify([configurationMock setAnalyticsCollectionEnabled:NO persistSetting:NO]);
+
+  defaultApp.dataCollectionDefaultState = FIRDataCollectionStateDefault;
+  OCMVerify([configurationMock setAnalyticsCollectionEnabled:YES persistSetting:NO]);
 }
 
 #pragma mark - Internal Methods